% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadNGouts.R
\name{GetApsimNGTable}
\alias{GetApsimNGTable}
\title{Read APSIM Next Generation output files.}
\usage{
GetApsimNGTable(dbLoc, table)
}
\arguments{
\item{dbLoc}{The location of the .db file.}

\item{table}{The name of the table to read.}
}
\description{
Reads APSIM NG .db files.
}
\details{
This function will read a single table from a given APSIM NG .db file.
}
\examples{
\dontrun{GetApsimNGTable("c:/outputs/Wheat.db", "Results")}
}
