% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetFunctions.R
\name{insertTavAmp}
\alias{insertTavAmp}
\title{Inserts Tav and Amp into a met object.}
\usage{
insertTavAmp(met)
}
\arguments{
\item{met}{A met file object where the tav and amp will be inserted.}
}
\value{
A met file object to which tav and amp has been added.
}
\description{
Amp is obtained by averaging the mean daily temperature of each month over 
the entire data period resulting in twelve mean temperatures, and then 
subtracting the minimum of these values from the maximum. Tav is obtained by 
averaging the twelve mean monthly temperatures.
}
\details{
The original documentation for the stand alone Tav_Amp program can be found at
\url{http://www.apsim.info/Portals/0/OtherProducts/tav_amp.pdf}.
}
\examples{
data(met)
insertTavAmp(met)
}

