% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsim.R
\name{kingsData}
\alias{kingsData}
\title{Weather data from Kingsthorpe, Queensland, Australia.}
\format{A data frame containing 365 rows and 10 columns.}
\source{
internal
}
\description{
A dataset containing a year of weather data from Kingsthorpe.
}
