% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetFunctions.R
\name{loadMet}
\alias{loadMet}
\title{Read an APSIM formatted met file.}
\usage{
loadMet(f)
}
\arguments{
\item{f}{The full path to the file to read.}
}
\value{
A metFile object containing the read met file.
}
\description{
This function reads in a previously formatted met file. Can be useful for
calculating tav and amp for a previously completed file or for reading into R
for further analysis.
}
\details{
For reading in other formats see the Importing External Data section in \code{\link{prepareMet}}.
See \code{\link{metFile}} for more information on the \code{met} object.
}
\examples{
\dontrun{loadMet("Weather.met")}
}
