% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsim.R
\name{met}
\alias{met}
\title{A complete metFile example using data from Dalby, Queensland, Australia.}
\format{See \code{\link{metFile}} for more information on the metFile class.}
\source{
internal
}
\description{
Three years of data from from Dalby including location coordinates, tav and
amp data and units for the weather data
}
