% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetFunctions.R
\docType{class}
\name{metFile-class}
\alias{metFile}
\alias{metFile-class}
\title{An S4 class used to store all information about a met file.}
\description{
An S4 class used to store all information about a met file.
}
\section{Slots}{

\describe{
\item{\code{const}}{A character vector containing constants that wilkl be written to the file. Format is 'variable = value'.}

\item{\code{lat}}{A length one numeric vector.}

\item{\code{lon}}{A length one numeric vector.}

\item{\code{tav}}{A length one numeric vector.}

\item{\code{amp}}{A length one numeric vector.}

\item{\code{units}}{A character vector containing the names of the columns in \code{data}.}

\item{\code{data}}{A data frame containing per day weather data.}
}}

