% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetFunctions.R
\name{writeMetFile}
\alias{writeMetFile}
\title{Write a metFile object to disk in APSIM met format.}
\usage{
writeMetFile(fileName, met)
}
\arguments{
\item{fileName}{The file name to write to.}

\item{met}{The metFile object to write.}
}
\description{
Takes a completed metFile object (generated via \code{\link{prepareMet}} or
\code{\link{loadMet}}) and writes it to disk.
}
\details{
Note the file will not be written if the met object is missing required
information such as latitude, tav or amp. Note that while longitude is required
in \code{\link{prepareMet}}, it is not strictly required by APSIM and thus is
optional here. However, it is best practise to include it so it will remain
mandatory when building a met file via \code{\link{prepareMet}}.
}
\examples{
data(Kingsthorpe)
newNames <-c("Date", "maxt", "mint", "rain", "evaporation", "radn", "vp", "Wind", "RH", "SVP")
units <- c("()", "(oC)", "(oC)", "(mm)", "(mm)", "(MJ/m^2/day)", "()", "()", "()", "()")
KingsMet <- prepareMet(kingsData, -27.48, 151.81, newNames = newNames, units = units)
writeMetFile(file.path(tempdir(), "KingsMet.met"), KingsMet)
}
