% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apsim.R
\docType{data}
\name{kingsData}
\alias{kingsData}
\title{Weather data from Kingsthorpe, Queensland, Australia.}
\format{A data frame containing 365 rows and 10 columns.}
\usage{
kingsData
}
\description{
A dataset containing a year of weather data from Kingsthorpe.
}
\keyword{datasets}

